#137 created endpoint for reading messages for current user (now for real)
